Refactor transaction processing system to use TransactionOld model
- Updated EmailEnricher, Neo4jExporter, TransactionProcessor, and other components to use TransactionOld instead of Transaction.
- Introduced Fetcher class for handling CSV file reading and transaction fetching.
- Removed OpenAiCategorizer and TransactionPipeline classes.
- Added TitleFormatter class for standardizing transaction titles using OpenAI.
- Replaced BankTransaction model with RawTransaction for better clarity.
- Adjusted service registrations in ServiceCollectionExtensions to reflect new architecture.
- Updated tests to accommodate changes in transaction models and processing logic.

+++ src/TransactionProcessingSystem/Configuration/AppSettings.cs
@@ -30,7 +30,7 @@ public record OpenAISettings
 
     [Required]
     [Range(0, 2)]
-    public required double Temperature { get; init; }
+    public required float Temperature { get; init; }
 }
 
 /// <summary>
